Refactor composer UI components to separate files
